integracion de simulacion de metodos de busqueda y otras cosas

+++ backend/src/main/java/com/backendeva/backend/dto/RespuestaDto.java
@@ -6,6 +6,6 @@ import lombok.Setter;
 @Getter
 @Setter
 public class RespuestaDto {
-    private Long preguntaId;
-    private Long respuestaId;
+    private Integer preguntaId;
+    private Integer respuestaId;
 }
